Moves all the header files in the include directory.
