+++ set_bounds.c
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "include/common.h"
 #include <cheri/cheric.h>
 #include <cheriintrin.h>
 #include <stdint.h>
